move google client to service
